Add new rule "no-invalid-terminating-properties"
Accidentally using undefined terminating properties leads to the expect
check not working as expected. For example,

expect(foo).to.be.falsy

results in a no-op that does not perform any check and is probably not
what the author intended. See https://github.com/chaijs/chai/issues/235
for a discussion of "truthy" and "falsy", including examples of people that
have run into this:
https://github.com/chaijs/chai/issues/235#issuecomment-97992643